+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -42,7 +42,7 @@ public class ReviewService {
         return getReviewById(review.getReviewId());
     }
 
-    public boolean delete(int id) {
+    public boolean delete(long id) {
         Review rew = reviewStorage.getReviewById(id)
                 .orElseThrow(() -> new NotFoundException("Отзыв не найден"));
 
@@ -52,23 +52,23 @@ public class ReviewService {
         return isDeleted;
     }
 
-    public Review getReviewById(int id) {
+    public Review getReviewById(long id) {
         return reviewStorage.getReviewById(id).
                 orElseThrow(() -> new NotFoundException("Отзыв не найден!"));
     }
 
-    public List<Review> getAllReviewsByParam(Optional<Integer> filmId, int count) {
+    public List<Review> getAllReviewsByParam(Optional<Long> filmId, long count) {
         if (filmId.isPresent()) {
             return reviewStorage.getReviewsByFilmId(filmId.get(), count);
         } else
             return reviewStorage.getAllReviewsByParam(count);
     }
 
-    public void putLikeOrDislikeToReview(int id, int userId, int vote) {
+    public void putLikeOrDislikeToReview(long id, long userId, long vote) {
         reviewStorage.putLikeOrDislikeToReview(id, userId, vote);
     }
 
-    public void deleteLikeOrDislikeToReview(int id, int userId, int vote) {
+    public void deleteLikeOrDislikeToReview(long id, long userId, long vote) {
         reviewStorage.deleteLikeOrDislikeToReview(id, userId, vote);
     }
 }
